requirements.txt
Flask>=2.0.0
Jinja2>=3.0.0
Werkzeug>=2.0.0
itsdangerous>=2.0.0
Pillow>=9.0.0